+++ src/main/java/leets/weeth/domain/board/application/mapper/PostMapper.java
@@ -19,20 +19,20 @@ public interface PostMapper {
 
     @Mappings({
             @Mapping(target = "id", ignore = true),
-            @Mapping(target = "createdAt",      ignore = true),
-            @Mapping(target = "modifiedAt",     ignore = true),
+            @Mapping(target = "createdAt", ignore = true),
+            @Mapping(target = "modifiedAt", ignore = true),
             @Mapping(target = "user", source = "user"),
             @Mapping(target = "part", expression = "java(user.getUserPart())"),
             @Mapping(target = "cardinalNumber", expression = "java(latest.getCardinalNumber())")
     })
     Post fromPostDto(PostDTO.Save dto, User user, Cardinal latest);
 
-    @Mapping(target = "id",             ignore = true)
-    @Mapping(target = "createdAt",       ignore = true)
-    @Mapping(target = "modifiedAt",      ignore = true)
-    @Mapping(target = "user",           source = "user")
+    @Mapping(target = "id", ignore = true)
+    @Mapping(target = "createdAt", ignore = true)
+    @Mapping(target = "modifiedAt", ignore = true)
+    @Mapping(target = "user", source = "user")
     @Mapping(target = "cardinalNumber", expression = "java(latest.getCardinalNumber())")
-    @Mapping(target = "category",       constant = "Education")
+    @Mapping(target = "category", constant = "Education")
     Post fromEducationDto(PostDTO.SaveEducation dto, User user, Cardinal latest);
 
     @Mappings({
